+++ environ.c
@@ -8,8 +8,8 @@
  */
 int _myenv(info_t *info)
 {
-    print_list_str(info->env);
-    return (0);
+	print_list_str(info->env);
+	return (0);
 }
 
 /**
@@ -21,17 +21,17 @@ int _myenv(info_t *info)
  */
 char *_getenv(info_t *info, const char *name)
 {
-    list_t *node = info->env;
-    char *p;
+	list_t *node = info->env;
+	char *p;
 
-    while (node)
-    {
-        p = starts_with(node->str, name);
-        if (p && *p)
-            return (p);
-        node = node->next;
-    }
-    return (NULL);
+	while (node)
+	{
+		p = starts_with(node->str, name);
+		if (p && *p)
+			return (p);
+		node = node->next;
+	}
+	return (NULL);
 }
 
 /**
@@ -43,14 +43,14 @@ char *_getenv(info_t *info, const char *name)
  */
 int _mysetenv(info_t *info)
 {
-    if (info->argc != 3)
-    {
-        _eputs("Incorrect number of arguements\n");
-        return (1);
-    }
-    if (_setenv(info, info->argv[1], info->argv[2]))
-        return (0);
-    return (1);
+	if (info->argc != 3)
+	{
+		_eputs("Incorrect number of arguements\n");
+		return (1);
+	}
+	if (_setenv(info, info->argv[1], info->argv[2]))
+		return (0);
+	return (1);
 }
 
 /**
@@ -61,17 +61,17 @@ int _mysetenv(info_t *info)
  */
 int _myunsetenv(info_t *info)
 {
-    int i;
+	int i;
 
-    if (info->argc == 1)
-    {
-        _eputs("Too few arguements.\n");
-        return (1);
-    }
-    for (i = 1; i <= info->argc; i++)
-        _unsetenv(info, info->argv[i]);
+	if (info->argc == 1)
+	{
+		_eputs("Too few arguements.\n");
+		return (1);
+	}
+	for (i = 1; i <= info->argc; i++)
+		_unsetenv(info, info->argv[i]);
 
-    return (0);
+	return (0);
 }
 
 /**
@@ -82,11 +82,11 @@ int _myunsetenv(info_t *info)
  */
 int populate_env_list(info_t *info)
 {
-    list_t *node = NULL;
-    size_t i;
+	list_t *node = NULL;
+	size_t i;
 
-    for (i = 0; environ[i]; i++)
-        add_node_end(&node, environ[i], 0);
-    info->env = node;
-    return (0);
+	for (i = 0; environ[i]; i++)
+		add_node_end(&node, environ[i], 0);
+	info->env = node;
+	return (0);
 }
