+++ exits.c
@@ -9,25 +9,25 @@
  */
 char *_strncpy(char *dest, char *src, int n)
 {
-    int i, j;
-    char *s = dest;
+	int i, j;
+	char *s = dest;
 
-    i = 0;
-    while (src[i] != '\0' && i < n - 1)
-    {
-        dest[i] = src[i];
-        i++;
-    }
-    if (i < n)
-    {
-        j = i;
-        while (j < n)
-        {
-            dest[j] = '\0';
-            j++;
-        }
-    }
-    return (s);
+	i = 0;
+	while (src[i] != '\0' && i < n - 1)
+	{
+		dest[i] = src[i];
+		i++;
+	}
+	if (i < n)
+	{
+		j = i;
+		while (j < n)
+		{
+			dest[j] = '\0';
+			j++;
+		}
+	}
+	return (s);
 }
 
 /**
@@ -39,22 +39,22 @@ char *_strncpy(char *dest, char *src, int n)
  */
 char *_strncat(char *dest, char *src, int n)
 {
-    int i, j;
-    char *s = dest;
+	int i, j;
+	char *s = dest;
 
-    i = 0;
-    j = 0;
-    while (dest[i] != '\0')
-        i++;
-    while (src[j] != '\0' && j < n)
-    {
-        dest[i] = src[j];
-        i++;
-        j++;
-    }
-    if (j < n)
-        dest[i] = '\0';
-    return (s);
+	i = 0;
+	j = 0;
+	while (dest[i] != '\0')
+		i++;
+	while (src[j] != '\0' && j < n)
+	{
+		dest[i] = src[j];
+		i++;
+		j++;
+	}
+	if (j < n)
+		dest[i] = '\0';
+	return (s);
 }
 
 /**
@@ -65,11 +65,10 @@ char *_strncat(char *dest, char *src, int n)
  */
 char *_strchr(char *s, char c)
 {
-    do
-    {
-        if (*s == c)
-            return (s);
-    } while (*s++ != '\0');
+	do {
+		if (*s == c)
+			return (s);
+	} while (*s++ != '\0');
 
-    return (NULL);
+	return (NULL);
 }
