+++ lists1.c
@@ -8,14 +8,14 @@
  */
 size_t list_len(const list_t *h)
 {
-    size_t i = 0;
+	size_t i = 0;
 
-    while (h)
-    {
-        h = h->next;
-        i++;
-    }
-    return (i);
+	while (h)
+	{
+		h = h->next;
+		i++;
+	}
+	return (i);
 }
 
 /**
@@ -26,33 +26,34 @@ size_t list_len(const list_t *h)
  */
 char **list_to_strings(list_t *head)
 {
-    list_t *node = head;
-    size_t i = list_len(head), j;
-    char **strs;
-    char *str;
+	list_t *node = head;
+	size_t i = list_len(head), j;
+	char **strs;
+	char *str;
 
-    if (!head || !i)
-        return (NULL);
-    strs = malloc(sizeof(char *) * (i + 1));
-    if (!strs)
-        return (NULL);
-    for (i = 0; node; node = node->next, i++)
-    {
-        str = malloc(_strlen(node->str) + 1);
-        if (!str)
-        {
-            for (j = 0; j < i; j++)
-                free(strs[j]);
-            free(strs);
-            return (NULL);
-        }
+	if (!head || !i)
+		return (NULL);
+	strs = malloc(sizeof(char *) * (i + 1));
+	if (!strs)
+		return (NULL);
+	for (i = 0; node; node = node->next, i++)
+	{
+		str = malloc(_strlen(node->str) + 1);
+		if (!str)
+		{
+			for (j = 0; j < i; j++)
+				free(strs[j]);
+			free(strs);
+			return (NULL);
+		}
 
-        str = _strcpy(str, node->str);
-        strs[i] = str;
-    }
-    strs[i] = NULL;
-    return (strs);
+		str = _strcpy(str, node->str);
+		strs[i] = str;
+	}
+	strs[i] = NULL;
+	return (strs);
 }
+
 
 /**
  * print_list - prints all elements of a list_t linked list
@@ -62,19 +63,19 @@ char **list_to_strings(list_t *head)
  */
 size_t print_list(const list_t *h)
 {
-    size_t i = 0;
+	size_t i = 0;
 
-    while (h)
-    {
-        _puts(convert_number(h->num, 10, 0));
-        _putchar(':');
-        _putchar(' ');
-        _puts(h->str ? h->str : "(nil)");
-        _puts("\n");
-        h = h->next;
-        i++;
-    }
-    return (i);
+	while (h)
+	{
+		_puts(convert_number(h->num, 10, 0));
+		_putchar(':');
+		_putchar(' ');
+		_puts(h->str ? h->str : "(nil)");
+		_puts("\n");
+		h = h->next;
+		i++;
+	}
+	return (i);
 }
 
 /**
@@ -87,16 +88,16 @@ size_t print_list(const list_t *h)
  */
 list_t *node_starts_with(list_t *node, char *prefix, char c)
 {
-    char *p = NULL;
+	char *p = NULL;
 
-    while (node)
-    {
-        p = starts_with(node->str, prefix);
-        if (p && ((c == -1) || (*p == c)))
-            return (node);
-        node = node->next;
-    }
-    return (NULL);
+	while (node)
+	{
+		p = starts_with(node->str, prefix);
+		if (p && ((c == -1) || (*p == c)))
+			return (node);
+		node = node->next;
+	}
+	return (NULL);
 }
 
 /**
@@ -108,14 +109,14 @@ list_t *node_starts_with(list_t *node, char *prefix, char c)
  */
 ssize_t get_node_index(list_t *head, list_t *node)
 {
-    size_t i = 0;
+	size_t i = 0;
 
-    while (head)
-    {
-        if (head == node)
-            return (i);
-        head = head->next;
-        i++;
-    }
-    return (-1);
+	while (head)
+	{
+		if (head == node)
+			return (i);
+		head = head->next;
+		i++;
+	}
+	return (-1);
 }
