+++ realloc.c
@@ -9,11 +9,11 @@
  */
 char *_memset(char *s, char b, unsigned int n)
 {
-    unsigned int i;
+	unsigned int i;
 
-    for (i = 0; i < n; i++)
-        s[i] = b;
-    return (s);
+	for (i = 0; i < n; i++)
+		s[i] = b;
+	return (s);
 }
 
 /**
@@ -22,13 +22,13 @@ char *_memset(char *s, char b, unsigned int n)
  */
 void ffree(char **pp)
 {
-    char **a = pp;
+	char **a = pp;
 
-    if (!pp)
-        return;
-    while (*pp)
-        free(*pp++);
-    free(a);
+	if (!pp)
+		return;
+	while (*pp)
+		free(*pp++);
+	free(a);
 }
 
 /**
@@ -41,22 +41,22 @@ void ffree(char **pp)
  */
 void *_realloc(void *ptr, unsigned int old_size, unsigned int new_size)
 {
-    char *p;
+	char *p;
 
-    if (!ptr)
-        return (malloc(new_size));
-    if (!new_size)
-        return (free(ptr), NULL);
-    if (new_size == old_size)
-        return (ptr);
+	if (!ptr)
+		return (malloc(new_size));
+	if (!new_size)
+		return (free(ptr), NULL);
+	if (new_size == old_size)
+		return (ptr);
 
-    p = malloc(new_size);
-    if (!p)
-        return (NULL);
+	p = malloc(new_size);
+	if (!p)
+		return (NULL);
 
-    old_size = old_size < new_size ? old_size : new_size;
-    while (old_size--)
-        p[old_size] = ((char *)ptr)[old_size];
-    free(ptr);
-    return (p);
+	old_size = old_size < new_size ? old_size : new_size;
+	while (old_size--)
+		p[old_size] = ((char *)ptr)[old_size];
+	free(ptr);
+	return (p);
 }
