+++ vars.c
@@ -10,29 +10,29 @@
  */
 int is_chain(info_t *info, char *buf, size_t *p)
 {
-    size_t j = *p;
+	size_t j = *p;
 
-    if (buf[j] == '|' && buf[j + 1] == '|')
-    {
-        buf[j] = 0;
-        j++;
-        info->cmd_buf_type = CMD_OR;
-    }
-    else if (buf[j] == '&' && buf[j + 1] == '&')
-    {
-        buf[j] = 0;
-        j++;
-        info->cmd_buf_type = CMD_AND;
-    }
-    else if (buf[j] == ';') /* found end of this command */
-    {
-        buf[j] = 0; /* replace semicolon with null */
-        info->cmd_buf_type = CMD_CHAIN;
-    }
-    else
-        return (0);
-    *p = j;
-    return (1);
+	if (buf[j] == '|' && buf[j + 1] == '|')
+	{
+		buf[j] = 0;
+		j++;
+		info->cmd_buf_type = CMD_OR;
+	}
+	else if (buf[j] == '&' && buf[j + 1] == '&')
+	{
+		buf[j] = 0;
+		j++;
+		info->cmd_buf_type = CMD_AND;
+	}
+	else if (buf[j] == ';') /* found end of this command */
+	{
+		buf[j] = 0; /* replace semicolon with null */
+		info->cmd_buf_type = CMD_CHAIN;
+	}
+	else
+		return (0);
+	*p = j;
+	return (1);
 }
 
 /**
@@ -47,26 +47,26 @@ int is_chain(info_t *info, char *buf, size_t *p)
  */
 void check_chain(info_t *info, char *buf, size_t *p, size_t i, size_t len)
 {
-    size_t j = *p;
+	size_t j = *p;
 
-    if (info->cmd_buf_type == CMD_AND)
-    {
-        if (info->status)
-        {
-            buf[i] = 0;
-            j = len;
-        }
-    }
-    if (info->cmd_buf_type == CMD_OR)
-    {
-        if (!info->status)
-        {
-            buf[i] = 0;
-            j = len;
-        }
-    }
+	if (info->cmd_buf_type == CMD_AND)
+	{
+		if (info->status)
+		{
+			buf[i] = 0;
+			j = len;
+		}
+	}
+	if (info->cmd_buf_type == CMD_OR)
+	{
+		if (!info->status)
+		{
+			buf[i] = 0;
+			j = len;
+		}
+	}
 
-    *p = j;
+	*p = j;
 }
 
 /**
@@ -77,25 +77,25 @@ void check_chain(info_t *info, char *buf, size_t *p, size_t i, size_t len)
  */
 int replace_alias(info_t *info)
 {
-    int i;
-    list_t *node;
-    char *p;
+	int i;
+	list_t *node;
+	char *p;
 
-    for (i = 0; i < 10; i++)
-    {
-        node = node_starts_with(info->alias, info->argv[0], '=');
-        if (!node)
-            return (0);
-        free(info->argv[0]);
-        p = _strchr(node->str, '=');
-        if (!p)
-            return (0);
-        p = _strdup(p + 1);
-        if (!p)
-            return (0);
-        info->argv[0] = p;
-    }
-    return (1);
+	for (i = 0; i < 10; i++)
+	{
+		node = node_starts_with(info->alias, info->argv[0], '=');
+		if (!node)
+			return (0);
+		free(info->argv[0]);
+		p = _strchr(node->str, '=');
+		if (!p)
+			return (0);
+		p = _strdup(p + 1);
+		if (!p)
+			return (0);
+		info->argv[0] = p;
+	}
+	return (1);
 }
 
 /**
@@ -106,36 +106,37 @@ int replace_alias(info_t *info)
  */
 int replace_vars(info_t *info)
 {
-    int i = 0;
-    list_t *node;
+	int i = 0;
+	list_t *node;
 
-    for (i = 0; info->argv[i]; i++)
-    {
-        if (info->argv[i][0] != '$' || !info->argv[i][1])
-            continue;
+	for (i = 0; info->argv[i]; i++)
+	{
+		if (info->argv[i][0] != '$' || !info->argv[i][1])
+			continue;
 
-        if (!_strcmp(info->argv[i], "$?"))
-        {
-            replace_string(&(info->argv[i]),
-                           _strdup(convert_number(info->status, 10, 0)));
-            continue;
-        }
-        if (!_strcmp(info->argv[i], "$$"))
-        {
-            replace_string(&(info->argv[i]),
-                           _strdup(convert_number(getpid(), 10, 0)));
-            continue;
-        }
-        node = node_starts_with(info->env, &info->argv[i][1], '=');
-        if (node)
-        {
-            replace_string(&(info->argv[i]),
-                           _strdup(_strchr(node->str, '=') + 1));
-            continue;
-        }
-        replace_string(&info->argv[i], _strdup(""));
-    }
-    return (0);
+		if (!_strcmp(info->argv[i], "$?"))
+		{
+			replace_string(&(info->argv[i]),
+				_strdup(convert_number(info->status, 10, 0)));
+			continue;
+		}
+		if (!_strcmp(info->argv[i], "$$"))
+		{
+			replace_string(&(info->argv[i]),
+				_strdup(convert_number(getpid(), 10, 0)));
+			continue;
+		}
+		node = node_starts_with(info->env, &info->argv[i][1], '=');
+		if (node)
+		{
+			replace_string(&(info->argv[i]),
+				_strdup(_strchr(node->str, '=') + 1));
+			continue;
+		}
+		replace_string(&info->argv[i], _strdup(""));
+
+	}
+	return (0);
 }
 
 /**
@@ -147,7 +148,7 @@ int replace_vars(info_t *info)
  */
 int replace_string(char **old, char *new)
 {
-    free(*old);
-    *old = new;
-    return (1);
+	free(*old);
+	*old = new;
+	return (1);
 }
